# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "Kitabxana idarəetməsi üçün CRUD REST API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "summary": "Bütün kitabları göstər (səhifələmə və çeşidləmə ilə)",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uğurlu"
                    }
                }
            },
            "post": {
                "summary": "Yeni kitab yarat",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "type": "string"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    },
                                    "published_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaradıldı"
                    },
                    "422": {
                        "description": "Doğrulama xətası"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "summary": "Tək kitabı göstər",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uğurlu"
                    },
                    "404": {
                        "description": "Tapılmadı"
                    }
                }
            },
            "put": {
                "summary": "Kitabı yenilə",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yeniləndi"
                    },
                    "404": {
                        "description": "Tapılmadı"
                    }
                }
            },
            "delete": {
                "summary": "Kitabı sil",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Silindi"
                    },
                    "404": {
                        "description": "Tapılmadı"
                    }
                }
            }
        }
    }
}